Initial commit
A Rust port of HashData[1].

Most of the test cases in this project have been hand crafted and
therefore may not represent actual valid hashes. Regexes used are as
close as possible to the versions found in HashData. Some of the regexes
have been made case insensitive to match found or generated example
hashes.

[1]: https://github.com/sam-b/hashdata